% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchaeologicalFloors-package.R
\docType{package}
\name{ArchaeologicalFloors}
\alias{ArchaeologicalFloors}
\alias{ArchaeologicalFloors-package}
\title{ArchaeologicalFloors}
\description{
Data and code to reproduce the results in Negre et al. (2016) which
implements a cost-based geostatistical analysis of chemical residues on
archaeological floors.
}
\references{
N. Negre, F. Muñoz, C. Lancelotti (2016). Geostatistical
  modelling of chemical residues on archaeological floors in the presence of
  barriers Journal of Archaeological Science 70 (2016) 91--101
}

