% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_helpers.R
\name{geom_india}
\alias{geom_india}
\title{Helper geom for the Jandhala case study}
\usage{
geom_india()
}
\description{
A specific ggplot2 geom with the structure of the units, the border and the
observations.
}
\details{
It keeps equal coordinates and removes axis labels.
}

