% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_helpers.R
\name{geom_variogram}
\alias{geom_variogram}
\title{Helper geom for variogram objects}
\usage{
geom_variogram(x, max.dist, scaled = FALSE, ...)
}
\arguments{
\item{x}{a \code{variomodel} object}

\item{max.dist}{miximum distance (x-axis) to compute and draw the variogram}

\item{scaled}{logical. If TRUE the total sill in the plot equals 1}

\item{...}{not used.

This function is an adaptation to \code{ggplot2} of the original function
\code{lines.variomodel} in \code{geoR}.}
}
\description{
Adds a \code{ggplot2} layer with a line plot of a variogram model.
}
\examples{
## Computing and plotting empirical variogram
library(geoRcb)
library(ggplot2)
vg <- variog(s100, max.dist = 1)
ggplot(vg)

## Estimating parameters by wheighted least squares
vgmdl <- variofit(vg, ini = c(1, .3), fix.nugget = TRUE)

## Adding fitted model to the empirical variogram
ggplot(vg) + geom_variogram(vgmdl)
}

