% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_helpers.R
\name{ggplot.variogram}
\alias{ggplot.variogram}
\title{Plotting empirical variograms with ggplot}
\usage{
\method{ggplot}{variogram}(data, ...)
}
\arguments{
\item{data}{a variogram object, as created for example with \code{variog}}

\item{...}{not used}
}
\description{
Defines a \code{variogram} method for \code{ggplot2}, for plotting empirical
variograms.
}

