% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchaeologicalFloors-package.R
\docType{data}
\name{Jandhala}
\alias{Jandhala}
\title{Jandhala}
\description{
Measurements (N = 70) of chemical residues taken in various points
distributed homogeneously over the floor in a domestic unit from Jandhala, a
village in the Patan district of North Gujarat, India. Specifically, Calcium
and a combination of several chemical elements (Calcium, Phosphorus,
Potassium, Magnesium and Strontium) which can account for the signature of
deposited food remains areas.
}
\details{
Variables include rectangular coordinates (x, y) of the observations in m, and the
measured values of Calcium (Ca) and food remains (foodremains).
}
\references{
N. Negre, F. Muñoz, C. Lancelotti (2016). Geostatistical
  modelling of chemical residues on archaeological floors in the presence of
  barriers Journal of Archaeological Science 70 (2016) 91--101
}

