% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv_helper.R
\name{krige.loocv}
\alias{krige.loocv}
\title{Leave-one-out Cross Validation Kriging}
\usage{
krige.loocv(method = c("std", "cst"), geodata, ddm)
}
\arguments{
\item{method}{string. Either \code{std} or \code{cst} for \emph{standard} Euclidean kriging or \emph{cost}-based kriging respectively.}

\item{geodata}{a geodata object}

\item{ddm}{cost-based distance matrices as in the output of \code{distmatGen}.}
}
\description{
Implements a Leave-one-out Cross Validation (LOOCV) procedure with either
Euclidean (\code{std}) or Cost-based (\code{cst}) kriging.
}

